Replace Python boolean representations with icons
Deploy to Bluemix and Free Tier Friendly are both columns defined by
booleans in YAML, which are currently being dumped into the HTML using
their Python representations ("True" and "False"). Instead of dumping
the representations, instead use the boolean value to select a
colored Unicode-based icon to represent the state (a green checkmark, or
a red X).